+++ typings/typings.d.ts
@@ -7,3 +7,4 @@
 /// <reference path="./custom/ambient/angular2/angular2.d.ts" />
 
 /// <reference path="./../node_modules/carbonldp/dist/bundles/carbon.d.ts" />
+/// <reference path="./../node_modules/angular2-carbonldp/dist/bundles/angular2-carbonldp.d.ts" />