+++ CHANGELOG.md
@@ -1,2 +1,11 @@
-# 1.0.0 (2016-03-17)
-* Initial release
+## 1.0.1 (2016-04-20) :maple_leaf:
+- Update `angular2` to 2.0.0-beta.15
+- Update `carbonldp` to 0.28.0
+- Integrate `angular2-carbonldp`
+- Configure !text imports meaning a file can be imported as text by appending !text to file name
+- Add style.css example for AppComponent.ts
+- Add gulp `build` and `serve:dist` task to show how a production build can be created
+- Add config files under `config/`
+
+## 1.0.0 (2016-03-17)
+- Initial release